refactor: split plugin policy parser helpers
